# docs/swagger.yaml
basePath: /api
definitions:
...
  customer.RegistrationDto:
    properties:
      age:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      waivers:
        items:
          $ref: '#/definitions/customer.WaiverSigningDto'
        type: array
    required:
    - age
    - first_name
    - last_name
    - waivers
    type: object
  customer.WaiverSigningDto:
    properties:
      is_waiver_signed:
        type: boolean
      waiver_url:
        type: string
    required:
    - waiver_url
    type: object
...
  entity.UserInfo:
    properties:
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      staffInfo:
        $ref: '#/definitions/staff.Details'
    type: object
...
  facility.CategoryRequestDto:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  facility.CategoryResponseDto:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  staff.Details:
    properties:
      createdAt:
        type: string
      isActive:
        type: boolean
      roleID:
        type: string
      roleName:
        type: string
      updatedAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user using Firebase token and returns a JWT token
        for the authenticated user
      parameters:
      - description: Firebase token for user verification
        in: header
        name: firebase_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User authenticated successfully
          schema:
            $ref: '#/definitions/entity.UserInfo'
        "400":
          description: 'Bad Request: Invalid Firebase token'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Authenticate a user and return a JWT token
      tags:
      - authentication
...
  /facility-categories:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all facility categories.
      produces:
      - application/json
      responses:
        "200":
          description: GetMemberships of facility categories retrieved successfully
          schema:
            items:
              $ref: '#/definitions/facility.CategoryResponseDto'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all facility categories
      tags:
      - facility-categories
    post:
      consumes:
      - application/json
      description: Registers a new facility category with the provided name.
      parameters:
      - description: Facility Category details
        in: body
        name: facility_category
        required: true
        schema:
          $ref: '#/definitions/facility.CategoryRequestDto'
      produces:
      - application/json
      responses:
        "201":
          description: Facility category created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid input'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new facility category
      tags:
      - facility-categories
  /facility-categories/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a facility category by its ID.
      parameters:
      - description: Facility Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: 'No Content: Facility category deleted successfully'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: Facility category not found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a facility category
      tags:
      - facility-categories
    get:
      consumes:
      - application/json
      description: Retrieves a facility category by its ID.
      parameters:
      - description: Facility Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Facility category retrieved successfully
          schema:
            $ref: '#/definitions/facility.CategoryResponseDto'
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: Facility category not found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a facility category by ID
      tags:
      - facility-categories
    put:
      consumes:
      - application/json
      description: Updates the details of an existing facility category by ID.
      parameters:
      - description: Facility Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated facility category details
        in: body
        name: facility_category
        required: true
        schema:
          $ref: '#/definitions/facility.CategoryRequestDto'
      produces:
      - application/json
      responses:
        "204":
          description: 'No Content: Facility category updated successfully'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid input'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'Not Found: Facility category not found'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update a facility category
      tags:
      - facility-categories
...
  /register/child:
    post:
      consumes:
      - application/json
      description: Registers a new child account using the provided details and associates
        it with the parent based on the Firebase authentication token.
      parameters:
      - description: Child account registration details
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/customer.RegistrationDto'
      - description: Firebase token for user verification
        in: header
        name: firebase_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Child account registered successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid input or missing Firebase token'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: Invalid Firebase token or insufficient permissions'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'Internal Server Error: Failed to register child account or
            associate with parent'
          schema:
            additionalProperties: true
            type: object
      summary: Register a new child account and associate it with the parent
      tags:
      - registration
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"